[path change] Page Components relocation within app directory
[+] Seperate TextureOverlay Component
