+++ app/Components/Cart/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { PageTransitionLayout } from "@/layouts/PageTransitionLayout";
-
+import {TextureOverlay} from "../TextureOverlay";
 const Cart = ({
   showCart,
   onClick,
@@ -10,13 +10,16 @@ const Cart = ({
 }) => {
   return (
     <PageTransitionLayout animate={showCart}>
-      <div className="grunge_overlay opacity-70"></div>
+      <TextureOverlay opacity={30} />
       <div className="w-full h-screen flex items-center justify-between">
         <div className="z-10 w-full flex flex-col items-center justify-between gap-5 mx-auto inset-x-0 max-w-max">
           <p className=" text-accent font-retro dropshadow text-lg">
             Whoooooooshhh! Empty.
           </p>
-          <button className="btn w-bw h-bh font-retro text-black" onClick={onClick}>
+          <button
+            className="btn w-bw h-bh font-retro text-black"
+            onClick={onClick}
+          >
             CLOSE
           </button>
         </div>
