+++ app/layout.tsx
@@ -1,4 +1,5 @@
 import type { Metadata } from "next";
+import { TextureOverlay } from "./Components/TextureOverlay";
 import "./globals.css";
 
 export const metadata: Metadata = {
@@ -14,8 +15,9 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className="bg-black">
-        <div className="grunge" />
-          <main>{children}</main>
+        {/* <div className="grunge" /> */}
+        <TextureOverlay opacity={50} />
+        <main>{children}</main>
       </body>
     </html>
   );
